Stlsupport - Using static polymorphism to simplify the code in the two writers. Redundant code removed from writer.h and prettywriter.h by having a new writerbase.h
